Rename: Fix Project Structure
클린 아키텍처 설계 & 사이즈 증가로 인한 프로젝트 구조 변경
