+++ IOS/BaekJoonProfile/BaekJoonProfile/Model/Network/SolvedACService.swift
@@ -45,9 +45,6 @@ struct SolvedACService{
         }
     }
     
-  
-    
-    
     private func isValidData(data: Data) -> DataState<Profile> {
         let decoder = JSONDecoder()
         
@@ -56,17 +53,4 @@ struct SolvedACService{
         return .Success(data: decodeData)
     }
 }
-enum ProfileError : Error {
-    case DefaultError
-    case ParsingError
-}
-extension ProfileError : LocalizedError {
-    public var errorDescription: String? {
-        switch self {
-        case .DefaultError:
-            return NSLocalizedString("네트워크나 아이디를 확인해주세요", comment: "기본 에러")
-        case .ParsingError:
-            return NSLocalizedString("프로필 파싱 에러", comment: "데이터 파싱 에러")
-        }
-    }
-}
+
